refactor!: simplify flag_evaluation namespace
