--- open_feature/flag_evaluation/flag_type.py
@@ -1,9 +1,0 @@
-from enum import Enum
-
-
-class FlagType(Enum):
-    BOOLEAN = "BOOLEAN"
-    STRING = "STRING"
-    OBJECT = "OBJECT"
-    FLOAT = "FLOAT"
-    INTEGER = "INTEGER"
