--- open_feature/flag_evaluation/reason.py
@@ -1,12 +1,0 @@
-from open_feature._backports.strenum import StrEnum
-
-
-class Reason(StrEnum):
-    CACHED = "CACHED"
-    DEFAULT = "DEFAULT"
-    DISABLED = "DISABLED"
-    ERROR = "ERROR"
-    STATIC = "STATIC"
-    SPLIT = "SPLIT"
-    TARGETING_MATCH = "TARGETING_MATCH"
-    UNKNOWN = "UNKNOWN"
