+++ open_feature/provider/in_memory_provider.py
@@ -4,8 +4,7 @@ import typing
 from open_feature._backports.strenum import StrEnum
 from open_feature.evaluation_context import EvaluationContext
 from open_feature.exception import ErrorCode
-from open_feature.flag_evaluation.reason import Reason
-from open_feature.flag_evaluation.resolution_details import FlagResolutionDetails
+from open_feature.flag_evaluation import FlagResolutionDetails, Reason
 from open_feature.hooks import Hook
 from open_feature.provider.metadata import Metadata
 from open_feature.provider.provider import AbstractProvider
