+++ open_feature/provider/no_op_provider.py
@@ -1,8 +1,7 @@
 import typing
 
 from open_feature.evaluation_context import EvaluationContext
-from open_feature.flag_evaluation.reason import Reason
-from open_feature.flag_evaluation.resolution_details import FlagResolutionDetails
+from open_feature.flag_evaluation import FlagResolutionDetails, Reason
 from open_feature.hooks import Hook
 from open_feature.provider.metadata import Metadata
 from open_feature.provider.no_op_metadata import NoOpMetadata
