+++ open_feature/provider/provider.py
@@ -2,7 +2,7 @@ import typing
 from abc import abstractmethod
 
 from open_feature.evaluation_context import EvaluationContext
-from open_feature.flag_evaluation.resolution_details import FlagResolutionDetails
+from open_feature.flag_evaluation import FlagResolutionDetails
 from open_feature.hooks import Hook
 from open_feature.provider.metadata import Metadata
