+++ tests/provider/test_in_memory_provider.py
@@ -1,8 +1,7 @@
 from numbers import Number
 
 from open_feature.exception import ErrorCode
-from open_feature.flag_evaluation.reason import Reason
-from open_feature.flag_evaluation.resolution_details import FlagResolutionDetails
+from open_feature.flag_evaluation import FlagResolutionDetails, Reason
 from open_feature.provider.in_memory_provider import InMemoryProvider, InMemoryFlag
